requirements.txt
Flask==2.2.2
Flask-SocketIO==5.3.2
PyYAML==6.0
eventlet==0.33.3
Werkzeug==2.2.2
Pillow==9.5.0
